docs(atomic-react): various atomic-react samples improvements
https://coveord.atlassian.net/browse/KIT-2195
