+++ packages/samples/atomic-react/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "atomic-react",
+  "name": "@coveo/atomic-react-samples",
   "version": "0.0.0",
   "description": "Samples with atomic-react",
   "private": true,
